pyproject.toml
[tool.poetry]
name = "in-delivery-backend"
version = "0.1.0"
description = ""
authors = ["Deverty <bychenkov.a.k@edu.mirea.ru>"]
readme = "README.md"
packages = [{ include = "src", from = "." }]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.105.0"
uvicorn = "^0.24.0.post1"
pydantic-settings = "^2.1.0"
pytest = "^7.4.3"
httpx = "^0.25.2"
redis = "^5.0.1"
fastapi-keycloak = "^1.0.10"
strawberry-graphql = "^0.216.1"
asyncpg = "^0.29.0"
fastapi-restful = "^0.5.0"


[tool.poetry.group.dev.dependencies]
poethepoet = "^0.24.4"
ipykernel = "^6.27.1"
asyncpg-stubs = "^0.29.1"
ruff = "^0.1.7"
mypy = "^1.7.1"
bandit = "^1.7.6"
typing-inspect = "^0.9.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
exclude = ["migrations"]

[tool.mypy]
plugins = "strawberry.ext.mypy_plugin"
disable_error_code = "type-abstract"
# do not check pep 695
